Add Hydration-test
