+++ src/Hydration.js
@@ -8,32 +8,36 @@ class Hydration {
     //2) filter through data to get all of it for the given user
     //3) calculate avg
     //4 return result
-    let userWaterConsumption = this.h20Data.filter(waterInfo => {
+    const userWaterConsumption = this.h20Data.filter(waterInfo => {
+      // console.log('this is user object', waterInfo)
       return waterInfo.userID === id;
     });
 
-    let totalOuncesConsumed = userWaterConsumption.reduce((totalOunces, waterInfo) => {
+    const totalOuncesConsumed = userWaterConsumption.reduce((totalOunces, waterInfo) => {
+      // console.log('this is totalOunces', totalOunces);
+      // console.log('userWaterConsumption', userWaterConsumption)
       return totalOunces += waterInfo.numOunces;
     }, 0);
 
-    let avgWaterConsumed = Math.round(totalOuncesConsumed / userWaterConsumption.length);
+    // let avgWaterConsumed = Math.round(totalOuncesConsumed / userWaterConsumption.length);
+    return Math.round(totalOuncesConsumed / userWaterConsumption.length);
 
-    console.log(`avg ounces consumed by user >>>> ${avgWaterConsumed}`);
-    return avgWaterConsumed;
+    // console.log(`avg ounces consumed by user >>>> ${avgWaterConsumed}`);
+    // return avgWaterConsumed;
   }
 
-  calculateDailyOunces(id) {
-    let userWaterConsumption = this.h20Data.filter(waterInfo => {
-      return waterInfo.userID === id;
-    });
-
-    let totalOuncesConsumed = userWaterConsumption.reduce((totalOunces, waterInfo) => {
-      return totalOunces += waterInfo.numOunces;
-    }, 0);
-
-    console.log(`daily ounces >>>> ${totalOuncesConsumed}`)
-    return totalOuncesConsumed;
-  }
+  // calculateDailyOunces(id) {
+  //   let userWaterConsumption = this.h20Data.filter(waterInfo => {
+  //     return waterInfo.userID === id;
+  //   });
+  //
+  //   let totalOuncesConsumed = userWaterConsumption.reduce((totalOunces, waterInfo) => {
+  //     return totalOunces += waterInfo.numOunces;
+  //   }, 0);
+  //
+  //   console.log(`daily ounces >>>> ${totalOuncesConsumed}`)
+  //   return totalOuncesConsumed;
+  // }
 }
 
 module.exports = Hydration;
