+++ test/User-test.js
@@ -2,13 +2,11 @@ import {
   expect
 } from 'chai';
 import User from '../src/User';
-import Hydration from '../src/Hydration';
 
 describe('User', () => {
   let user1;
   let user2;
   let user3;
-  let hydrationInfo;
 
   beforeEach(() => {
     user1 = new User({
@@ -31,25 +29,9 @@ describe('User', () => {
       "friends": [9, 18, 24, 19]
     })
 
-    let hydrationData = [
-      {userID: 1, date: '2019/06/15', numOunces: 37},
-      {userID: 1, date: '2019/06/16', numOunces: 42},
-      {userID: 1, date: '2019/06/17', numOunces: 54}
-    ];
-
-    hydrationInfo = new Hydration(hydrationData);
   })
 
   it('should be able to return a users first name', () => {
     expect(user2.getFirstName()).to.equal('Jarvis')
-  })
-
-  // it('should be able to calculate total ounces consumed in a specific day', () => {
-  //   const totalOunces = hydrationInfo.calculateDailyOunces();
-  //   expect(totalOunces).to.equal(133);
-  // })
-
-  it('should return how much water the user consumed on average', () => {
-    expect(hydrationInfo.getAvgOuncesPerDay(1)).to.equal(44);
   })
 })
